requirements.txt
grpcio
grpcio-tools
redis
fastapi
uvicorn
websockets
jinja2